requirements.txt
telepot==12.7
xlrd==1.2.0
XlsxWriter==1.1.2
urllib3==1.24.1
pymongo==3.7.2
python-telegram-bot==11.1.0
matplotlib==3.0.2
